% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBlankStartStateVector.R
\name{createBlankStartStateVector}
\alias{createBlankStartStateVector}
\title{Create a blank (named) state vector containing zero values.}
\usage{
createBlankStartStateVector(params, types)
}
\arguments{
\item{\code{params}}{A named parameter vector.  Must contain a parameter called "gamma_shape" for the time taken (assumed Gamma distributed with rate parameter equal to 1) for development from an egg to adult.}

\item{\code{types}}{is a character vector containing the types of insects.  One of these should be named "Wld" for the wildtype.}
}
\value{
The function returns a named numeric vector for the model.
}
\description{
\code{createBlankStartStateVector} uses the information about the types of insects and the parameters to construct a blank state vector.  This can then be populated with the correct numbers of each type to initiate a run of a model.
}
