% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantiles.R
\name{plotQuantiles}
\alias{plotQuantiles}
\title{A function for visualising an object of type \code{trajectoryQuantileList}.}
\usage{
plotQuantiles(quantileList, times, name, col, xlab = "Time",
  ylab = "Number", main = "")
}
\arguments{
\item{\code{quantileList}}{is an object of class trajectoryQuantileList created using the function called trajectoryQuantiles.}

\item{\code{times}}{is a numberic vector containing the times that the trajectories in \code{trajectoryQuantileList} correspond to.}

\item{\code{name}}{is the name of the state variable to be plotted (e.g. "Wld_m").}

\item{\code{col}}{A colour (preferably without added transparency) to use for plotting the trajectories.  Transparency will be added in the function.}
}
\value{
The function returns a visualisation of temporal variability in a trajectory highlighting the 2.5th, 25th, 50th, 75th and 97.5th percetiles.
}
\description{
\code{fade} provides the name of a semi-transparent R colour after passing the name of a base colour or colours (e.g. "blue") and the degree of opacity desired.
}
