% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIIT_metapopulation.R
\name{simulateIIT_metapopulation}
\alias{simulateIIT_metapopulation}
\title{Simulate an IIT Metapopulation program trajectory as a Markov Population Process.}
\usage{
simulateIIT_metapopulation(paramsList, blockImmigrationRates, Wld_m,
  Wld_f_Unmated, Wld_f_Wld, stochasticInitial = TRUE, numReleased = NULL,
  ratioReleased = NULL, releaseMixture, contaminationProb, propTypes,
  releaseTimes, maxTime, initStateList = NULL, maxSize = 1e+06)
}
\arguments{
\item{\code{paramList}}{A list where each item is a named parameter vector containing the model parameters for a metapopulation patch.  This must include the birth rate "lambda"; the death rates of males and females "mu_m" and "mu_f"; the parameters "gamma_shape" and "gamma_rate" that govern the gamma distribution for time taken for immatures to develop from egg to adult; "K_eq" which is the number of adults we expect at equilibrium; "N_max" which is the ceiling on the number of immatures that can exist in the population at any point in time; and a Fried's index (denoted by the prefix "c_") for each type of insect in the population (e.g. "c_Wld" is the Fried's index for Wildtype males).}

\item{\code{Wld_m}}{a numeric vector of length equal to the number of rows or columns in blockImmigrationRates and containing the number of wildtype males that are expected at time zero.}

\item{\code{Wld_f_Unmated}}{a numeric vector of length equal to the number of rows or columns in blockImmigrationRates and containing the number of unmated wildtype females that are expected in a pure wildtype population at equilibrium.}

\item{\code{Wld_f_Wld}}{a numeric vector of length equal to the number of rows or columns in blockImmigrationRates and containing the number of wildtype mated wildtype females that are expected in a pure wildtype population at equilibrium.}

\item{\code{stochasticInitial}}{should the initial numbers be sampled treating \code{propTypes} as probabilities.}

\item{\code{numReleased}}{is the number of insects to be released at each each release time.  This should be a matrix, with a row for each metapopulation patch and with columns equal to the number of release times.}

\item{\code{ratioReleased}}{is the ratio of insects to be released at each each release time (instead of numReleased), relative to the number of males in the wild.  This should be a numeric vector of length equal to the number of release times.}

\item{\code{releaseMixture}}{a numeric vector containing the proportions of each insect type in the released mixture.  The vector should be named according to the insect types.}

\item{\code{contaminationProb}}{a numeric vector containing the probabilities that the a released insect is a female rather than a male for each insect type released.}

\item{\code{propTypes}}{is the proportions of each type in the initial population.  This should be a numeric vector with names correcponding to the types of insects.}

\item{\code{releaseTimes}}{is a numeric vector containing the times at which the insects are released.}

\item{\code{maxTime}}{is the maximum time to run the simulation for.  Simulations where the entire population goes extinct will be terminated earlier.}

\item{\code{initStateList}}{a list containing numeric vectors that acts as an override allowing you to specify the initial state vector directly.}

\item{\code{maxSize}}{provides control over the blocksize used to extend the matrices used to store simulation outputs.}
}
\value{
The function returns a named list containing a numeric vector of times ("t") and a matrix of corresponding states ("states").
}
\description{
\code{simulateIIT} simualtes a Markov Population Process model for a Wolbachia IIT release and returns a (stochastic) simulated trajectory for the system.
}
