% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationToDaily.R
\name{simulationToDaily}
\alias{simulationToDaily}
\title{Convert a simulated birth-death process trajectory into a multivariate timeseries at specified time points.}
\usage{
simulationToDaily(times, states, newTimes, relevantStateNames)
}
\arguments{
\item{\code{times}}{a numeric vector containing the event occurence times in a simulated birth-death process.}

\item{\code{states}}{a numeric matrix containing (as rows) the state vectors of the simulated birth-death process at the times given in \code{times}.  The column names of the matrix should contain the state names of each element in the state vector.}

\item{\code{newTimes}}{a numeric vector containing the times at which the constructed multivariate time series should be created (can be at regular or irregular intervals).}

\item{\code{relevantStateNames}}{a character vector containing the names of the state variables that should be included in the returned in the multivariate timeseries. These names should be elements in the set of column names in \code{states}.}
}
\value{
a matrix object with rows corresponding to each of the time points in the input \code{times}.
}
\description{
\code{simulationToDaily} creates a matrix whose rows are the state of a simulated birth-death process at specific points in time.  The function is useful for turning a simulation into a multivariate timeseries on a regular timestep for example.
}
