% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryQuantiles.R
\name{trajectoryQuantiles}
\alias{trajectoryQuantiles}
\title{Create an object of class \code{trajectoryQuantileList} that summarises the variability in an ensemble of simulated trajectories.}
\usage{
trajectoryQuantiles(regSimArray, relevantStateNames)
}
\arguments{
\item{\code{regSimArray}}{an array object that stores an ensemble of simulated trajectories that share a regular timestep.  The array must have three dimensions: the first corresponds to the number of simulations in the ensemble; the second corresponds to the number of time-points in the simulated trajectories (these must have been regularised to a common set of times); and the third being the number of states in the state vector of the simulated process.}

\item{\code{relevantStateNames}}{a character vector containing the names of the elements in the state vector for \code{regSimArray} (i.e. the state names).}
}
\value{
a matrix object with rows corresponding to each of the time points in the input \code{times}.
}
\description{
\code{trajectoryQuantiles} returns an object of class \code{trajectoryQuantileList} that contains named elements that summarise, at each point in time, the 2.5th, 25th, 50th, 75th and 97.5th percentiles of an ensemble of trajectories.
}
